+++ src/controllers/UserController.ts
@@ -1,11 +1,10 @@
 /**
- * @file Dispatcher for user related requests.
+ * @file Controller for user related requests.
  * @author Sebastian Gadzinski
  */
-import express from 'express';
 import Result from '../classes/Result';
 import User from '../models/User';
-import UserService from '../services/UserService';
+import { SecurityService, UserService } from '../services';
 
 class UserController {
   public getAll(req: any, res: any) {
@@ -59,6 +58,76 @@ class UserController {
       .catch((errorMessage: Error) =>
         res.send(new Result({ message: errorMessage.message, success: false }))
       );
+  }
+
+  public async getProfile(req: any, res: any) {
+    try {
+      // if not admin
+      const userId = req?.params?.id;
+
+      // If userId is a thing and its not the user who is on check for admin
+      if (userId && !req?.user?.data.roles.includes('admin')) {
+        await SecurityService.accessDenied(req.ip);
+      }
+
+      // If no userId then retrun the users own
+      let user: any = {};
+      const queryUserId = userId ?? req.user.data.id;
+      const selectionQuery: any = {
+        _id: 0,
+        name: '$fullName',
+        email: 1,
+        phoneNumber: 1
+      };
+      if (req?.user?.data.roles.includes('admin')) {
+        selectionQuery.emailConfirmed = 1;
+      }
+
+      user = await User.findOne({ _id: queryUserId }, selectionQuery);
+
+      if (!user?.phoneNumber) {
+        user.phoneNumber = '';
+      }
+
+      res.send(new Result({ data: user, success: true }));
+    } catch (err) {
+      res.send(new Result({ message: err.message, success: false }));
+    }
+  }
+
+  public async saveProfile(req: any, res: any) {
+    try {
+      // if not admin
+      const userId = req?.body?.userId;
+
+      // If userId is a thing and its not the user who is on check for admin
+      if (userId && !req?.user?.data.roles.includes('admin')) {
+        await SecurityService.accessDenied(req.ip);
+      }
+
+      const setQuery: any = {
+        fullName: req.body.user.name, // New name value
+        phoneNumber: req.body.user.phoneNumber
+      };
+
+      if (req?.user?.data.roles.includes('admin')) {
+        setQuery.email = req.body.user.email;
+        setQuery.emailConfirmed = req.body.user.emailConfirmed;
+      }
+
+      const queryUserId = userId ?? req.user.data.id;
+
+      await User.updateOne(
+        { _id: queryUserId }, // Filter to match the user to update
+        {
+          $set: setQuery
+        }
+      );
+
+      res.send(new Result({ success: true }));
+    } catch (err) {
+      res.send(new Result({ message: err.message, success: false }));
+    }
   }
 }
 export default new UserController();
