+++ src/controllers/index.ts
@@ -5,7 +5,7 @@
 export { default as AppController } from './AppController';
 export { default as AuthenticationController } from './AuthenticationController';
 export { default as BrowseController } from './BrowseController';
-export { default as DataController } from './DataController';
+export { default as DataController } from './HomeController';
 export { default as MeetingController } from './MeetingController';
 export { default as UserController } from './UserController';
 export { default as VMController } from './VMController';
