+++ src/routes/index.ts
@@ -7,7 +7,7 @@ export { default as appRoutes } from './app';
 export { default as authenticationRoutes } from './authentication';
 export { default as browseRoutes } from './browse';
 export { default as classRoutes } from './class';
-export { default as dataRoutes } from './data';
+export { default as homeRoutes } from './home';
 export { default as meetingRoutes } from './meeting';
 export { default as userRoutes } from './user';
 export { default as vmRoutes } from './vm';
