+++ src/routes/user.ts
@@ -2,7 +2,7 @@
  * @file Defines all user related route patterns.
  * @author Sebastian Gadzinski
  */
-import express, { response } from 'express';
+import express from 'express';
 
 import UserController from '../controllers/UserController';
 import {
@@ -11,6 +11,12 @@ import {
 } from '../middleware/authenticationMiddleware';
 
 const router = express.Router({});
+
+// Profile
+router.get('/profile', isAuthenticated, UserController.getProfile);
+router.get('/profile/:id', isAuthenticated, UserController.getProfile);
+
+router.post('/profile/save', isAuthenticated, UserController.saveProfile);
 
 // Create
 router.post(
@@ -25,7 +31,12 @@ router.get('/', isAuthenticated, UserController.getAll);
 router.get('/:id', isAuthenticated, UserController.get);
 
 // Update
-router.put('/:id', isAuthenticated, UserController.update);
+router.put(
+  '/:id',
+  isAuthenticated,
+  isUser({ minRole: 'admin' }),
+  UserController.update
+);
 
 // Delete
 router.delete(
